pyproject.toml
[project]
name = "agent-runner"
version = "2.0.0"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.48",
    "fastmcp>=3.1.0",
    "a2a-sdk[http-server]>=0.3.24",
    "PyJWT[crypto]",
    "uvicorn",
    "google-cloud-firestore",
    "google-cloud-pubsub",
    "httpx",
    "pyyaml",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
